test(types): вынести общие хелперы tests/maxo/types в conftest
